# Faster xdr reader