--- a/LICENSE
+++ b/LICENSE
@@ -1,7 +1,6 @@
 MIT License
 
-Copyright (c) 2018 Carlos Ramos Carreño
-Copyright (c) 2023 CSC - IT Center for Science Ltd.
+Copyright (c) 2018 Rdata developers.
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal