This PR adds faster reader for files in xdr format. Full arrays are read directly with numpy instead of reading element by element. As a positive side effect, deprecated xdrlib isn't needed anymore.

Related to #31. I have cherry-picked and rebased commits related to xdr reader improvements to this PR. There are also some structural changes that are open for discussion, for example, xdr reader is moved to `rdata/io/xdr.py` to simplify separation between different readers like (upcoming) `rdata/io/ascii.py`. @vnmabus Could you review?