@trossi 2023-12-08
Here is approximate timing data for reference:

| Array size (MiB) | Time to read before this PR (s) | Time to read with this PR (s) |
| ---: | ---: | ---: |
| 16 | 1.2 | 0.3 |
| 32 | 2.2 | 0.3 |
| 64 | 4.2 | 0.3 |
| 128 | 8.0 | 0.4 |
| 256 | 18.5 | 0.5 |
| 512 | 39.4 | 0.8 |
| 1024 | 77.3 | 1.5 |

This data was created with the following (in bash):
1. Generate test data (without compression to skip time spent in decompression): `for i in {1..7}; do n=$(( 2 ** $i )); Rscript -e "saveRDS(runif(n=$n*1024**2), file='array_$i.rds', compress=FALSE)"; done`
2. Read and measure time: `for i in {1..7}; do echo $i; time -p python -c "from rdata.parser import parse_file; parse_file('array_$i.rds')"; done`