@vnmabus 2024-01-25
> Here is approximate timing data for reference:

Just to let you know: I added your example (but limited to 5 iterations) as an [asv benchmark](https://asv.readthedocs.io/en/v0.6.1/index.html) to the package, to check for future performance regressions. I also added a new `testing` module (currently undocumented) to retrieve and execute R snippets from strings, so that each test can have its associated R snippet for creating the data, instead of a big script for all.